+++ NEWS.md
@@ -2,6 +2,7 @@
 
 ## Minor tweaks
 * Added internal functions and checks for the input arguments.
+* Added `p_exp` (-log10 p-values) to the output of `cpma` and `fct`.
 
 # ctar 0.2.0